backend/requirements.txt
flask
flask-cors
chromadb
pdfplumber
requests
python-dotenv
